+++ libraries/matchers.rb
@@ -47,7 +47,7 @@ if defined?(ChefSpec)
     oneview_server_profile:             standard_actions,
     oneview_storage_pool:               [:add_if_missing, :remove],
     oneview_storage_system:             [:add, :remove, :edit_credentials, :add_if_missing],
-    oneview_switch:                     [:remove, :none],
+    oneview_switch:                     [:remove, :none, :patch],
     oneview_unmanaged_device:           [:add, :remove, :add_if_missing],
     oneview_uplink_set:                 standard_actions,
     oneview_volume:                     standard_actions + [:create_snapshot, :delete_snapshot],